backend/requirements.txt
absl-py==2.3.1
adal==1.2.7
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==3.7.1
applicationinsights==0.11.10
argcomplete==3.5.3
asgiref==3.11.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
azure-appconfiguration==1.7.1
azure-batch==15.0.0b2
azure-cli==2.72.0
azure-cli-core==2.72.0
azure-cli-telemetry==1.1.0
azure-common==1.1.28
azure-core==1.34.0
azure-cosmos==3.2.0
azure-data-tables==12.4.0
azure-datalake-store==1.0.0a0
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys==4.9.0b3
azure-keyvault-secrets==4.7.0
azure-mgmt-advisor==9.0.0
azure-mgmt-apimanagement==4.0.0
azure-mgmt-appconfiguration==3.1.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-applicationinsights==1.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-batch==17.3.0
azure-mgmt-batchai==7.0.0b1
azure-mgmt-billing==6.0.0
azure-mgmt-botservice==2.0.0
azure-mgmt-cdn==12.0.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-compute==34.1.0
azure-mgmt-containerinstance==10.2.0b1
azure-mgmt-containerregistry==11.0.0
azure-mgmt-containerservice==35.0.0
azure-mgmt-core==1.5.0
azure-mgmt-cosmosdb==9.7.0
azure-mgmt-databoxedge==1.0.0
azure-mgmt-datalake-store==1.1.0b1
azure-mgmt-datamigration==10.0.0
azure-mgmt-dns==8.0.0
azure-mgmt-eventgrid==10.2.0b2
azure-mgmt-eventhub==10.1.0
azure-mgmt-extendedlocation==1.0.0b2
azure-mgmt-hdinsight==9.0.0b3
azure-mgmt-imagebuilder==1.3.0
azure-mgmt-iotcentral==10.0.0b2
azure-mgmt-iothub==3.0.0
azure-mgmt-iothubprovisioningservices==1.1.0
azure-mgmt-keyvault==11.0.0
azure-mgmt-loganalytics==13.0.0b4
azure-mgmt-managementgroups==1.0.0
azure-mgmt-maps==2.0.0
azure-mgmt-marketplaceordering==1.1.0
azure-mgmt-media==9.0.0
azure-mgmt-monitor==5.0.1
azure-mgmt-msi==7.0.0
azure-mgmt-mysqlflexibleservers==1.0.0b3
azure-mgmt-netapp==10.1.0
azure-mgmt-policyinsights==1.1.0b4
azure-mgmt-postgresqlflexibleservers==1.1.0b2
azure-mgmt-privatedns==1.0.0
azure-mgmt-rdbms==10.2.0b17
azure-mgmt-recoveryservices==3.0.0
azure-mgmt-recoveryservicesbackup==9.1.0
azure-mgmt-redhatopenshift==1.5.0
azure-mgmt-redis==14.5.0
azure-mgmt-resource==23.1.1
azure-mgmt-search==9.1.0
azure-mgmt-security==6.0.0
azure-mgmt-servicebus==8.2.1
azure-mgmt-servicefabric==2.1.0
azure-mgmt-servicefabricmanagedclusters==2.1.0b1
azure-mgmt-servicelinker==1.2.0b3
azure-mgmt-signalr==2.0.0b2
azure-mgmt-sql==4.0.0b21
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==22.1.0
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==7.3.1
azure-monitor-query==1.2.0
azure-multiapi-storage==1.4.0
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.20.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.7.0
bcrypt==4.3.0
beautifulsoup4==4.14.2
billiard==4.2.4
blinker==1.9.0
cachetools==5.5.0
celery==5.3.4
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.14
decorator==5.1.1
Deprecated==1.2.18
distro==1.9.0
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
executing==2.2.0
fabric==3.2.2
filelock==3.20.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.55.3
fsspec==2025.9.0
future==1.0.0
gast==0.7.0
geocoder==1.38.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==21.2.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
inflection==0.5.1
invoke==2.2.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
javaproperties==0.5.2
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
jsondiff==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kanren==0.2.3
keras==3.13.0
kiwisolver==1.4.7
knack==0.11.0
kombu==5.6.2
libclang==18.1.1
lightgbm==4.6.0
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
microsoft-security-utilities-secret-masker==1.0.0b4
ml_dtypes==0.5.4
mpi4py==4.0.1
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
multipledispatch==1.0.0
namex==0.1.0
narwhals==1.14.3
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.3.0
opt_einsum==3.4.0
optree==0.18.0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parso==0.8.4
Pillow==10.0.1
pkginfo==1.12.1.2
platformdirs==4.3.8
portalocker==2.10.1
prompt_toolkit==3.0.51
proto-plus==1.27.1
protobuf==5.29.0
psutil==7.0.0
pure_eval==0.2.3
py-deviceid==0.1.1
pyarrow==18.1.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycomposefile==0.0.32
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyGithub==1.59.1
Pygments==2.18.0
PyJWT==2.10.1
pymsalruntime==0.17.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.2.0
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.4.0
ratelim==0.1.6
redis==4.6.0
referencing==0.35.1
regex==2025.10.23
reportlab==4.0.7
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.21.0
safetensors==0.6.2
scikit-learn==1.4.2
scipy==1.15.2
scp==0.13.6
seaborn==0.13.2
semver==2.13.0
setuptools==80.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
sqlparse==0.5.5
sshtunnel==0.1.5
stack-data==0.6.3
streamlit==1.40.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
textblob==0.19.0
threadpoolctl==3.6.0
tkintermapview==1.29
tkinterweb==3.24.8
tokenizers==0.22.1
toml==0.10.2
toolz==1.0.0
torch==2.8.0
torchvision==0.23.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unification==0.2.2
uritemplate==4.2.0
urllib3==2.2.3
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.3.3
Werkzeug==3.0.1
wheel==0.45.1
whitenoise==6.6.0
wrapt==1.17.2
xmltodict==0.14.2
